feat:crud completed

+++ gift_list_api/src/main/java/br/com/project/gift_list_api/GiftRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GiftRepository extends JpaRepository<Gift,Float> {
+public interface GiftRepository extends JpaRepository<Gift,Long> {
 }
